Fixed the OscarSerializer field_mapping overing, was raising a TypeError
OscarSerializer was extending the field_mapping dictionary by passsing
in another dict as a kwargs. At least on my system this was raising a
TypeError stating that keys must be strings (at least on my settup).

I've updated the method used to be more explicit (still does the same
thing under the hood).