fixes get_api_class not finding modules registered with OSCARAPI_OVERRIDE_MODULES
When get_api_class or get_api_classes is imported in a file that is parsed
during the django settings configuration, OSCARAPI_OVERRIDE_MODULES could be
pinned to the wrong value.

get_api_class now reads the value from settings when it is called instead of
statically defining it in the module. It is now safe to::

    from oscarapi.utils.loading import get_api_class, get_api_classes

In files that are parsed diring django setting configuration, avoiding ugly
constructs such as::

    def load_api_stuff():
        from oscarapi.utils.loading import get_api_class, get_api_classes
        api_class = get_api_class(...)
        # eulgh